Instances to separate module
